Crate tlsf

A Two Level Segregated Fit (TLSF) allocator optimized for memory-constrained systems

Features

Bounded execution time for the alloc and dealloc operations (WCET analysis friendly)

Optimized memory footprint (fixed cost)

Only 4 bytes of metadata (overhead) per allocation (variable cost)

The "size" of the allocator can be configured at compile time (pay for what you use)

Example

#![no_main]
#![no_std]

use core::{
    alloc::Layout,
    sync::atomic::{AtomicUsize, Ordering},
};

// see https://github.com/japaric/alloc-many
// or use the (still) unstable `alloc` crate
use alloc_many_collections::vec::Vec;
// or use the unstable `#[global_allocator]` / `#[alloc_error_handler]` features
use alloc_many::{oom, Alloc};
use cortex_m::asm;
use cortex_m_rt::entry;
use panic_halt as _; // panic handler
use spin::Mutex; // don't use this in interrupts because it can deadlock
use tlsf::{Tlsf, MAX_BLOCK_SIZE};

static ALLOC: Mutex<Tlsf> = Mutex::new(Tlsf::new());

// An allocator singleton
struct A;

// or use the `#[alloc_many::allocator]` attribute
unsafe impl Alloc for A {
    unsafe fn alloc(layout: Layout) -> *mut u8 {
        ALLOC.lock().alloc(layout)
    }

    unsafe fn dealloc(ptr: *mut u8, _: Layout) {
        ALLOC.lock().dealloc(ptr)
    }

    unsafe fn realloc(
        ptr: *mut u8,
        layout: Layout,
        new_size: usize,
    ) -> *mut u8 {
        ALLOC.lock().realloc(ptr, layout, new_size)
    }

    unsafe fn alloc_zeroed(layout: Layout) -> *mut u8 {
        ALLOC.lock().alloc_zeroed(layout)
    }
}

#[repr(align(4))]
struct Aligned<T>(T);

#[entry]
fn main() -> ! {
    // NOTE(Aligned) align this buffer to `tlsf::ALIGN_SIZE` bytes to avoid padding
    // NOTE(SIZE) use a size multiple of `MAX_BLOCK_SIZE` to maximize coalescing
    // NOTE(bss) initialize this to all zeros to put `M` in the `.bss` linker
    //           section and close to `tlsf::ANCHOR`
    static mut M: Aligned<[u8; MAX_BLOCK_SIZE as usize]> =
        Aligned([0; MAX_BLOCK_SIZE as usize]);

    ALLOC.lock().grow(&mut M.0);

    // allocate a vector in allocator `A`
    let mut xs: Vec<A, _> = Vec::new();
    loop {
        // push until we run out of memory (see `oom` function below)
        xs.push(X.fetch_add(1, Ordering::Relaxed) as u8);
    }
}

static X: AtomicUsize = AtomicUsize::new(0);

#[oom] // instead of the unstable `#[alloc_error_handler]`
fn oom(_: Layout) -> ! {
    // check `X` at this point
    asm::bkpt();

    loop {}
}

Cargo features

The TLSF allocator has two main parameters associated to it: the First Level Index (FLI) and the Second Level Index (SLI). In a nutshell: the SLI has an effect on fragmentation and the FLI controls how big the memory blocks managed by the allocator can get. See 2 for the full details.

The greater the value of the SLI the lesser the fragmentation. This crate hard codes the value of SLI to the recommended value of 4.

The bigger the value of FLI the bigger the memory blocks you can request from the allocator. However, increasing the FLI also increases the memory footprint of the allocator. This crate lets you pick the the value of FLI through Cargo features named: FLI6, FLI7, ..., FLI16. The table below shows the footprint of Tlsf and the maximum block size you can request from it for every supported value of FLI.

FLI	size_of(Tlsf)	MAX_REQUEST_SIZE	MAX_BLOCK_SIZE

6	36	60	60

7	70	120	124

8	104	240	252

9	138	480	508

10	172	960	1,020

11	206	1,920	2,044

12	240	3,840	4,092

13	274	7,680	8,188

14	308	15,360	16,380

15	342	30,720	32,764

16	376	61,440	65,532

All sizes are in bytes

MAX_REQUEST_SIZE is the maximum block size with alignment of ALIGN_SIZE bytes (or less) that can be requested from the allocator. Requesting a bigger size will immediately result in OOM. If you request a block with alignment greater than ALIGN_SIZE bytes then MAX_REQUEST_SIZE will effectively be lower because the allocator may need to insert padding to satisfy the alignment requirement.

MAX_BLOCK_SIZE is the maximum size of any individual memory block managed by the allocator. This number is not the maximum amount of memory the allocator can manage but you should consider it when you use Tlsf.grow because the allocator will split the given slice into blocks of MAX_BLOCK_SIZE bytes if it's too big. These initial blocks -- they are more like regions -- will never be coalesced by the allocator; also, the allocator will not coalesce blocks from different regions even if they are next to each other.

Limitations

To minimize the memory footprint of the allocator this implementation internally uses 16-bit offsets instead of pointers (halves memory footprint on a 32-bit machine). These offsets are computed from the address of the tlsf::ANCHOR variable. This decision limits the location of the memory blocks that the allocator can manage to those whose address are within ±(1 << 17) bytes (±128 KiBi) of the tlsf::ANCHOR variable, will be located in the output .bss linker section.

Benchmarks

Platform: ARM Cortex-M3 @ 8 MHz (zero Flash wait cycles).

Compilation flags: opt-level=3, lto=true, codegen-units=1

Cargo features: +FLI8

Code	Time

Tlsf.alloc

- no split	83

- split	159

- split & pad	294

Tlsf.dealloc

- no merge	80

- merge next	203

- merge previous	201

- merge both	263

All times are in core clock cycles (1 core clock cycle = 125 ns)

Minimum Supported Rust Version (MSRV)

This crate is guaranteed to compile on stable Rust 1.34 and up. It might compile on older versions but that may change in any new patch release.

References

Implementation of a constant-time dynamic storage allocator

TLSF: a New Dynamic Memory Allocator for Real-Time Systems

Structs

Anchor
A 1-byte struct that's ALIGN_SIZE-bytes aligned

Tlsf
The Two Level Segregated Fit (TLSF) allocator

Constants

ALIGN_SIZE
All block sizes are multiple of this number; this number is also the minimum alignment of all allocations

FLI
First Level Index (FLI)

MAX_BLOCK_SIZE
Maximum block size that can be managed by the allocator

MAX_REQUEST_SIZE
Maximum block size that can be requested from the allocator

SLI
Second Level Index (FLI)

Statics

ANCHOR
All internal offsets are relative to this "anchor"