Struct tlsf::Tlsf

pub struct Tlsf { /* fields omitted */ }

The Two Level Segregated Fit (TLSF) allocator

Methods

impl Tlsf

pub const fn new() -> Self

Constructs a new, empty Tlsf allocator

pub fn extend(&mut self, memory: &'static mut [u8]) -> usize

Adds a memory chunk to the Tlsf memory pool

This method returns the amount of usable memory in bytes

Note that the given memory chunk may not be fully utilized due to alignment requirements. Align memory to ALIGN_SIZE bytes to avoid any padding.

Also note that memory must be located within 128 KiBi of the tlsf::ANCHOR variable, which will be located in the output .bss linker section. Passing a memory chunk outside this address range will result in it being discarded and the grow call will return 0.

Finally note that the allocator will split memory into blocks of MAX_BLOCK_SIZE bytes if it's bigger than MAX_BLOCK_SIZE. These initial blocks, or regions, will never be coalesced at runtime; also, blocks from different regions will not be coalesced even if they are next to each other.

To get the most of memory use the initialization method shown in the example

pub unsafe fn alloc(&mut self, layout: Layout) -> Result<NonNull<u8>, ()>

Returns a pointer meeting the size and alignment guarantees of layout.

See core::alloc::Alloc.alloc for more details.

This method has a bounded execution time, meaning that it will complete within some time T regardless of the input and the state of the allocator -- the implementation of this method contains no loops, recursion or panicking branches.

pub unsafe fn dealloc(&mut self, ptr: NonNull<u8>)

Deallocate the memory referenced by ptr.

See core::alloc::Alloc.dealloc for more details

This method has a bounded execution time, meaning that it will complete within some time T regardless of the input and the state of the allocator -- the implementation of this method contains no loops, recursion or panicking branches.

pub unsafe fn grow_in_place(
&mut self,
ptr: NonNull<u8>,
new_size: usize
) -> Result<(), ()>

Attempts to extend the allocation referenced by ptr to fit new_size.

See core::alloc::Alloc.grow_in_place for more details

This method has a bounded execution time, meaning that it will complete within some time T regardless of the input and the state of the allocator -- the implementation of this method contains no loops, recursion or panicking branches.

pub unsafe fn shrink_in_place(&mut self, ptr: NonNull<u8>, new_size: usize)

Attempts to shrink the allocation referenced by ptr to fit new_size.

See core::alloc::Alloc.shrink_in_place for more details

This method has a bounded execution time, meaning that it will complete within some time T regardless of the input and the state of the allocator -- the implementation of this method contains no loops, recursion or panicking branches.

pub unsafe fn realloc(
&mut self,
ptr: NonNull<u8>,
layout: Layout,
new_size: usize
) -> Result<NonNull<u8>, ()>

Returns a pointer suitable for holding data described by a new layout with layout’s alignment and a size given by new_size. To accomplish this, this may extend or shrink the allocation referenced by ptr to fit the new layout.

See core::alloc::Alloc.realloc for more details

This method does not have a bounded execution time. It may complete in O(N) time where N = layout.size()